refactor: 상대경로를 절대경로로 수정 #73
