--- src/app/concert/[id]/_shared/services/query-key.ts
@@ -1,10 +1,0 @@
-import { GetConcertDetailRequest } from '@/app/concert/[id]/_shared/services/type';
-
-const queryKey = {
-  getConcertDetail: (request?: GetConcertDetailRequest) => [
-    'getConcertDetail',
-    request,
-  ],
-};
-
-export default queryKey;
