+++ src/app/history/_shared/components/agent-form-card/agent-form-card.tsx
@@ -2,15 +2,15 @@
 import Image from 'next/image';
 import Link from 'next/link';
 
-import { useGetConcertDetail } from '@/app/concert/[id]/_shared/services/query';
+import { useGetConcertDetail } from '@/app/concert/[id]/_shared/services/concert/query';
+import AcceptModal from '@/app/history/_shared/components/modal/common-modal';
+import RejectedModal from '@/app/history/_shared/components/modal/rejected-modal/rejected-modal';
 import Button from '@/shared/components/button/functional-button/functional-button';
 import { MODAL_ID } from '@/shared/components/modal/modal-constants';
 import { useModal } from '@/shared/components/modal/use-modal';
 import { Form } from '@/shared/types';
 
 import styles from './agent-form-card.module.scss';
-import AcceptModal from '../modal/common-modal';
-import RejectedModal from '../modal/rejected-modal/rejected-modal';
 
 interface FormCardProps {
   formItem: Form;
