--- src/shared/components/bottom-sheet/bottom-sheet.module.scss
@@ -1,81 +1,0 @@
-@use '@/styles/index' as *;
-
-// BottomSheet 스타일
-.container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  position: fixed;
-
-  width: 100%;
-  max-width: var(--max-width);
-  height: 220px;
-  bottom: 0;
-
-  border-radius: 24px 24px 0 0;
-  background: var(--white);
-  z-index: 50;
-  padding: 0 24px;
-
-  transform: translateY(100%);
-  transition: transform 0.5s ease-in-out;
-
-  &.open {
-    transform: translateY(0);
-  }
-
-  .icon {
-    width: 57px;
-    height: 4px;
-    border-radius: 7px;
-    background: var(--gray-4);
-    flex-shrink: 0;
-
-    &:active {
-      cursor: grabbing;
-    }
-  }
-
-  .upper_container {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-
-    .info {
-      @include text-sm;
-      color: var(--textColor-less);
-      margin-top: 8px;
-    }
-
-    .title_container {
-      display: flex;
-      align-items: flex-end;
-      width: 100%;
-      margin-top: 4px;
-
-      .title {
-        @include title-md;
-        color: var(--textColor-main);
-      }
-
-      .star {
-        @include text-md;
-        color: var(--textColor-main);
-        margin-right: 4px;
-      }
-    }
-  }
-
-  .button_container {
-    display: flex;
-    justify-content: space-between;
-    width: 100%;
-    margin-top: 20px;
-    gap: 16px;
-
-    a {
-      width: 100%;
-    }
-  }
-}
