--- src/shared/components/bottom-sheet/bottom-sheet.tsx
@@ -1,144 +1,0 @@
-'use client';
-
-import { useEffect, useState } from 'react';
-
-import Link from 'next/link';
-
-import { StarIcon } from '@/assets/icons';
-import Button from '@/shared/components/button/functional-button/functional-button';
-import { useCheckDuplicateForm } from '@/shared/services/bottom-sheet/mutation';
-import { Concert, TicketOpenType } from '@/shared/types';
-
-import styles from './bottom-sheet.module.scss';
-
-interface BottomSheetProps {
-  onClose: () => void;
-  isOpen: boolean;
-  concertItem: Concert | undefined;
-  concertId: string;
-  agentId: string;
-}
-
-const BottomSheet = ({
-  isOpen,
-  onClose,
-  concertItem,
-  concertId,
-  agentId,
-}: BottomSheetProps) => {
-  const { ticketOpenDateInfoResponses } = concertItem ?? {};
-
-  const preOpen = ticketOpenDateInfoResponses?.find(
-    (info) => info.ticketOpenType === 'PRE_OPEN',
-  );
-  const generalOpen = ticketOpenDateInfoResponses?.find(
-    (info) => info.ticketOpenType === 'GENERAL_OPEN',
-  );
-
-  // 각 타입별로 중복 여부 관리
-  const [isDuplicateMap, setIsDuplicateMap] = useState<
-    Record<TicketOpenType, boolean>
-  >({ PRE_OPEN: false, GENERAL_OPEN: false });
-
-  //선예매/일반예매를 분리해서 중복확인
-  const { mutate: checkPreOpen } = useCheckDuplicateForm();
-  const { mutate: checkGeneralOpen } = useCheckDuplicateForm();
-
-  useEffect(() => {
-    if (!isOpen || !agentId || !concertId) return;
-
-    if (preOpen) {
-      checkPreOpen(
-        { agentId, concertId, ticketOpenType: 'PRE_OPEN' },
-        {
-          onSuccess: (res) =>
-            setIsDuplicateMap((prev) => ({
-              ...prev,
-              PRE_OPEN: res as boolean,
-            })),
-          onError: () =>
-            setIsDuplicateMap((prev) => ({
-              ...prev,
-              PRE_OPEN: false,
-            })),
-        },
-      );
-    }
-
-    if (generalOpen) {
-      checkGeneralOpen(
-        { agentId, concertId, ticketOpenType: 'GENERAL_OPEN' },
-        {
-          onSuccess: (res) =>
-            setIsDuplicateMap((prev) => ({
-              ...prev,
-              GENERAL_OPEN: res as boolean,
-            })),
-          onError: () =>
-            setIsDuplicateMap((prev) => ({
-              ...prev,
-              GENERAL_OPEN: false,
-            })),
-        },
-      );
-    }
-  }, [
-    isOpen,
-    agentId,
-    concertId,
-    preOpen,
-    generalOpen,
-    checkGeneralOpen,
-    checkPreOpen,
-  ]);
-
-  return (
-    <div className={`${styles.container} ${isOpen ? styles.open : ''}`}>
-      <div className={styles.icon} onClick={onClose} />
-      <div className={styles.upper_container}>
-        <span className={styles.info}>프로필 보기</span>
-
-        <div className={styles.title_container}>
-          <span className={styles.title}>의문의 티켓터</span>
-          <StarIcon width={20} height={20} />
-          <span className={styles.star}>4.6</span>
-          <span className={styles.info}>(12)</span>
-        </div>
-
-        <span className={styles.info}>한 줄 소개를 작성해주세요.</span>
-      </div>
-
-      <div className={styles.button_container}>
-        {preOpen &&
-          (isDuplicateMap.PRE_OPEN ? (
-            <Button size="large" variant="fill-disabled" onClick={onClose}>
-              신청된 요청입니다
-            </Button>
-          ) : (
-            <Link href={`/concert/form/${concertId}?ticketOpenType=PRE_OPEN`}>
-              <Button size="large" variant="fill" onClick={onClose}>
-                선예매 요청하기
-              </Button>
-            </Link>
-          ))}
-
-        {generalOpen &&
-          (isDuplicateMap.GENERAL_OPEN ? (
-            <Button size="large" variant="fill-disabled" onClick={onClose}>
-              신청된 요청입니다
-            </Button>
-          ) : (
-            <Link
-              href={`/concert/form/${concertId}?ticketOpenType=GENERAL_OPEN`}
-            >
-              <Button size="large" variant="border" onClick={onClose}>
-                일반예매 요청하기
-              </Button>
-            </Link>
-          ))}
-      </div>
-    </div>
-  );
-};
-
-export default BottomSheet;
